pyproject.toml
[project]
name = "agentcore-data-pipeline"
version = "1.0.0"
description = "AgentCore Data Analysis Pipeline with Strands Graph Orchestration"
requires-python = ">=3.10"
dependencies = [
    # Strands and AgentCore
    "strands-agents>=1.10.0",
    "bedrock-agentcore>=0.1.0",

    # FastAPI for AgentCore endpoints
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",

    # AWS Integration
    "boto3>=1.28.0",
    "botocore>=1.31.0",

    # Core Data Analysis (minimal)
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # MCP integration
    "mcp>=1.0.0",

    # Streamlit Dashboard
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["agent"]
packages = []

[dependency-groups]
dev = []